BENTO-2277 Extract existing components to bento-core
