+++ src/bento-core/GlobalSearch/SearchPage/components/cards/aboutCard.js
