+++ src/bento-core/GlobalSearch/SearchPage/components/cards/fileCard.js
