+++ src/bento-core/GlobalSearch/SearchPage/components/cards/noAccessCard.js
