+++ src/bento-core/GlobalSearch/SearchPage/components/cards/sampleCard.js
