+++ src/bento-core/GlobalSearch/SearchPage/components/cards/studyCard.js
