+++ src/bento-core/GlobalSearch/SearchPage/components/component.js
