+++ src/bento-core/GlobalSearch/SearchPage/components/propertyItem.js
