+++ src/bento-core/GlobalSearch/SearchPage/components/tabs/publicTabView.js
@@ -1,18 +1,15 @@
 import React from 'react';
 import { withStyles } from '@material-ui/core';
 import {
-  SEARCH_PAGE_RESULT_ABOUT, SEARCH_PAGE_RESULT_FILES,
-  SEARCH_PAGE_RESULT_MODEL, SEARCH_PAGE_RESULT_PROGRAM,
+  SEARCH_PAGE_RESULT_ABOUT_PUBLIC,
+  SEARCH_PAGE_RESULT_FILES, SEARCH_PAGE_RESULT_MODEL, SEARCH_PAGE_RESULT_PROGRAM_PUBLIC,
   SEARCH_PAGE_RESULT_SAMPLES, SEARCH_PAGE_RESULT_STUDIES,
-  SEARCH_PAGE_RESULT_SUBJECTS,
-} from '../../../../bento/search';
+  SEARCH_PAGE_RESULT_SUBJECTS, SEARCH_PUBLIC,
+} from '../../../../../bento/search';
 import styles from '../../styles';
 import SearchViewTabs from './searchViewTabs';
 
-const allCount = (searchResults) => (searchResults.subject_count
-  + searchResults.sample_count + searchResults.program_count
-  + searchResults.study_count + searchResults.file_count
-  + searchResults.model_count + searchResults.about_count);
+const allCount = (searchResults) => (searchResults.about_count || 0);
 
 const getTabProperties = (classes, searchResults, allLabel) => [
   {
@@ -22,7 +19,7 @@ const getTabProperties = (classes, searchResults, allLabel) => [
       root: classes.buttonRoot,
       wrapper: classes.allTab,
     },
-    queryForApi: SEARCH_PAGE_RESULT_SUBJECTS,
+    queryForApi: SEARCH_PUBLIC,
     count: allCount(searchResults) || 0,
     value: '1',
   },
@@ -35,7 +32,7 @@ const getTabProperties = (classes, searchResults, allLabel) => [
     },
     queryForApi: SEARCH_PAGE_RESULT_SUBJECTS,
     count: searchResults.subject_count || 0,
-    value: '2',
+    value: 'inactive-2',
   },
   {
     name: 'Samples',
@@ -46,7 +43,7 @@ const getTabProperties = (classes, searchResults, allLabel) => [
     },
     queryForApi: SEARCH_PAGE_RESULT_SAMPLES,
     count: searchResults.sample_count || 0,
-    value: '3',
+    value: 'inactive-3',
   },
   {
     name: 'Files',
@@ -57,7 +54,7 @@ const getTabProperties = (classes, searchResults, allLabel) => [
     },
     queryForApi: SEARCH_PAGE_RESULT_FILES,
     count: searchResults.file_count || 0,
-    value: '4',
+    value: 'inactive-4',
   },
   {
     name: 'Programs',
@@ -66,9 +63,9 @@ const getTabProperties = (classes, searchResults, allLabel) => [
       root: classes.buttonRoot,
       wrapper: classes.programTab,
     },
-    queryForApi: SEARCH_PAGE_RESULT_PROGRAM,
+    queryForApi: SEARCH_PAGE_RESULT_PROGRAM_PUBLIC,
     count: searchResults.program_count || 0,
-    value: '5',
+    value: 'inactive-5',
   },
   {
     name: 'Studies',
@@ -79,7 +76,7 @@ const getTabProperties = (classes, searchResults, allLabel) => [
     },
     queryForApi: SEARCH_PAGE_RESULT_STUDIES,
     count: searchResults.study_count || 0,
-    value: '6',
+    value: 'inactive-6',
   },
   {
     name: 'Data Model',
@@ -90,7 +87,7 @@ const getTabProperties = (classes, searchResults, allLabel) => [
     },
     queryForApi: SEARCH_PAGE_RESULT_MODEL,
     count: searchResults.model_count || 0,
-    value: '7',
+    value: 'inactive-7',
   },
   {
     name: 'About',
@@ -99,17 +96,18 @@ const getTabProperties = (classes, searchResults, allLabel) => [
       root: classes.buttonRoot,
       wrapper: classes.aboutTab,
     },
-    queryForApi: SEARCH_PAGE_RESULT_ABOUT,
+    queryForApi: SEARCH_PAGE_RESULT_ABOUT_PUBLIC,
     count: searchResults.about_count || 0,
     value: '8',
   },
 ];
 
-const PrivateTabView = ({
+const PublicTabView = ({
   classes, options, tab, searchText,
 }) => {
   const { searchResults } = options;
   const tabProperties = getTabProperties(classes, searchResults);
+
   const AllLabel = () => (
     <div>
       <img
@@ -125,12 +123,12 @@ const PrivateTabView = ({
     <SearchViewTabs
       AllLabel={AllLabel}
       classes={classes}
-      tab={tab}
-      isPublic={false}
-      searchText={searchText}
+      isPublic
       options={{ ...options, properties: tabProperties }}
+      tab={tab}
+      searchText={searchText}
     />
   );
 };
 
-export default withStyles(styles, { withTheme: true })(PrivateTabView);
+export default withStyles(styles, { withTheme: true })(PublicTabView);
