+++ src/bento-core/GlobalSearch/SearchPage/components/tabs/searchViewTabs.js
