+++ src/bento-core/GlobalSearch/SearchPage/searchViewController.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 import SearchView from './searchView';
-import { PUBLIC_ACCESS } from '../../bento/siteWideConfig';
-import accessLevelTypes from '../../utils/enums';
+import { PUBLIC_ACCESS } from '../../../bento/siteWideConfig';
+import accessLevelTypes from '../../../utils/enums';
 
 const SearchViewContainer = ({ match }) => {
   const isSignedIn = useSelector((state) => state.login.isSignedIn);
