+++ src/bento-core/GlobalSearch/SearchPage/styles.js
