+++ src/components/Layout/LayoutView.js
@@ -20,8 +20,8 @@ import GraphqlClient from '../GraphqlClient/GraphqlView';
 import fileCentricCart from '../../pages/fileCentricCart/cartController';
 // import JBrowse from '../JBrowse/JBrowseView';
 import JBrowseDetail from '../../pages/jbrowseDetail/jbrowseDetailController';
-import GlobalSearch from '../../pages/search/searchView';
-import GlobalSearchController from '../../pages/search/searchViewController';
+import GlobalSearch from '../../bento-core/GlobalSearch/SearchPage/searchView';
+import GlobalSearchController from '../../bento-core/GlobalSearch/SearchPage/searchViewController';
 import adminController from '../../pages/admin/adminController';
 import reviewRequestController from '../../pages/admin/reviewPendingDAR/reviewRequestController';
 import Login from '../../pages/accessManagment/login';
